sof-apl-da7219: mark DMIC as device 99
Somehow there seems to be a bug in the topology, the DMIC cannot
be device 0 or the device crashes and reboots, with no oops or log.

Mark it as device 99 so that we keep in mind this is a complete hack
to deliver a working solution, we'll make it pretty later.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>